chore: Refactor search component
